+++ Package.swift
@@ -17,9 +17,9 @@ let package = Package(
         // Targets can depend on other targets in this package and products from dependencies.
         .binaryTarget(
             name: "BevyIosGamecenterRust",
-            // path: "BevyIosGamecenterRust.xcframework"),
-            url: "https://github.com/rustunit/bevy_ios_gamecenter/releases/download/rs-0.2.0/BevyIosGamecenterRust.xcframework.zip",
-            checksum: "65c2155602c88efa61b7acd156d382b7c0e88b0cd8fb05b851a84b81448df854"),
+             path: "BevyIosGamecenterRust.xcframework"),
+//            url: "https://github.com/rustunit/bevy_ios_gamecenter/releases/download/rs-0.2.0/BevyIosGamecenterRust.xcframework.zip",
+//            checksum: "65c2155602c88efa61b7acd156d382b7c0e88b0cd8fb05b851a84b81448df854"),
         .target(
             name: "bevy_ios_gamecenter",
             dependencies: ["BevyIosGamecenterRust"]),