+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+* new `resolve_conflicting_games` method (#8)
+* new `ConflictingSaveGames` and `ResolvedConflicts` events
+
+### Changed
+* new container type `IosGCSaveGames` now used in `IosGCSaveGamesResponse`
+
 ## [0.2.0] - 2024-07-12
 
 ### Changed